+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+* Make it easier to use different Selenium images
+* Update default Selenium version to ``4.25.0``
+* Updated dependencies
+
 # 1.0.2
 * Updated dependencies